Switch to stdint types

+++ exercises/grains/src/example.c
@@ -2,13 +2,13 @@
 
 #define NUMBER_OF_SQUARES   (64)
 
-unsigned long square(int index) {
-  return 1ul << (index - 1);
+uint64_t square(uint8_t index) {
+  return (uint64_t)1 << (index - 1);
 }
 
-unsigned long total(void) {
-  unsigned long result = 0;
-  int i;
+uint64_t total(void) {
+  uint64_t result = 0;
+  uint8_t i;
 
   for(i = 1; i <= NUMBER_OF_SQUARES; i++) {
     result += square(i);
